% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labourcost.R
\name{labourcost}
\alias{labourcost}
\title{Calculate Labour Cost of Wire Harness}
\usage{
labourcost(data_sga, wirecost, connectorcost, remainingpartcost, location)
}
\arguments{
\item{data_sga}{data-set of SGA rates}

\item{wirecost}{raw wire cost}

\item{connectorcost}{total connector & terminal cost}

\item{remainingpartcost}{total remaining part cost}

\item{location}{procurement location}
}
\value{
labour cost of wireharness
}
\description{
Calculate Labour Cost of Wire Harness
}
