% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manufacturingcost.R
\name{manufacturingcost}
\alias{manufacturingcost}
\title{Calculate manufacturing cost of wireharness}
\usage{
manufacturingcost(
  data_sga,
  wirecost,
  connectorcost,
  remainingpartcost,
  labour,
  location
)
}
\arguments{
\item{data_sga}{data-set of SGA rates}

\item{wirecost}{raw wire cost}

\item{connectorcost}{total connector and terminal cost}

\item{remainingpartcost}{total remaining part cost}

\item{labour}{labour cost}

\item{location}{procurement location}
}
\value{
manufacturing cost of wireharness
}
\description{
Calculate manufacturing cost of wireharness
}
