% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterscost.R
\name{parameterscost}
\alias{parameterscost}
\title{Calculate Other Parameters of Wire harness}
\usage{
parameterscost(data_sga, labour, manufacturing, location, parameters)
}
\arguments{
\item{data_sga}{data-set SGA rates}

\item{labour}{labour cost}

\item{manufacturing}{manufacturing cost}

\item{location}{procurement location}

\item{parameters}{SGA, ICC, Packing, Profit}
}
\value{
the parameter cost of wireharness
}
\description{
Calculate Other Parameters of Wire harness
}
