% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remainingpartcost.R
\name{remainingpartcost}
\alias{remainingpartcost}
\title{Calculate Approx Remaining Part Cost of Wire Harness}
\usage{
remainingpartcost(data_sga, wirecost, connectorcost, location)
}
\arguments{
\item{data_sga}{data object of SGA}

\item{wirecost}{raw wire cost}

\item{connectorcost}{total connector & terminal cost}

\item{location}{procurement location}
}
\value{
total remaining cost
}
\description{
Calculate Approx Remaining Part Cost of Wire Harness
}
