% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wirecost.R
\name{wirecost}
\alias{wirecost}
\title{Calculate Raw Wire Cost}
\usage{
wirecost(data_gauge, wirelengthh, Awg)
}
\arguments{
\item{data_gauge}{data-set of gauge rates}

\item{wirelengthh}{numeric string which mentions the total wire length}

\item{Awg}{numeric string which mentions Gauge of wire}
}
\value{
Raw Wire Cost (mostly in dollars)
}
\description{
Calculate Raw Wire Cost
}
