% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wirelength.R
\name{wirelength}
\alias{wirelength}
\title{Calculate Wire Length}
\usage{
wirelength(copperlength)
}
\arguments{
\item{copperlength}{numeric string where copper length should be in mm.}
}
\value{
wirelength in meter
}
\description{
Calculate Wire Length
}
